moved asset, user, and activity search methods to SearchController
